novel
Novels is the perfect app for you! Download novels now and re-discover the fun of reading.
With the application of novels, you can read online and download thousands of Arabic novels and you can store them directly in the memory of the device to be able to open them easily at any time without the need to connect to the Internet. Novels also give you many features and features:

- The application is completely free and you can read it online and download many Arabic novels and translator